refactor: reorganize project structure
- Add __init__.py files for easier imports
- Moved crud functions to app/crud/user.py
- Moved models to app/models/user.py
- Moved schemas to app/schemas/user.py
- Moved routers to app/routers/user.py
- Updated imports to match new structure
