+++ app/crud/__init__.py
@@ -1,0 +1,7 @@
+from .user import (
+    get_user_by_email,
+    create_user,
+    authenticate_user,
+    get_user,
+    get_users
+)
