+++ app/dependencies/__init__.py
@@ -1,0 +1,1 @@
+from .auth import get_current_user, oauth2_scheme
