+++ app/routers/__init__.py
@@ -1,0 +1,5 @@
+from fastapi import APIRouter
+from . import user
+
+router = APIRouter()
+router.include_router(user.router, tags=["users"])
