+++ app/routers/user.py
@@ -1,0 +1,73 @@
+from datetime import timedelta
+from fastapi import APIRouter, Depends, HTTPException, status
+from app import crud, models, schemas
+from app.auth import ACCESS_TOKEN_EXPIRE_MINUTES, create_access_token
+from app.database import get_db
+from app.dependencies import get_current_user
+from sqlalchemy.orm import Session
+
+router = APIRouter()
+
+@router.post("/register", response_model=schemas.Message)
+def register_user(user: schemas.UserRegister, db: Session = Depends(get_db)):
+    if crud.get_user_by_email(db, user.email):
+        raise HTTPException(status_code=400, detail="Email already registered")
+    created_user = crud.create_user(db, user)
+    if not created_user:
+        raise HTTPException(status_code=400, detail="Passwords do not match")
+    return {"message": "Account successfully registered."}
+
+@router.post("/login", response_model=schemas.Token)
+def login_user(user: schemas.UserLogin, db: Session = Depends(get_db)):
+    db_user = crud.authenticate_user(db, user.email, user.password)
+    if not db_user:
+        raise HTTPException(status_code=401, detail="Invalid login credentials")
+    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    access_token = create_access_token(data={"user_id": db_user.id}, expires_delta=access_token_expires)
+    return {"access_token": access_token, "token_type": "bearer"}
+
+
+@router.get("/users/me", response_model=schemas.UserRead)
+def read_me(current_user: models.User = Depends(get_current_user)):
+    return current_user
+
+@router.get("/users/{user_id}", response_model=schemas.UserRead)
+def read_user(user_id: int, db: Session = Depends(get_db)):
+    user = crud.get_user(db, user_id)
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    return user
+
+@router.get("/users", response_model=list[schemas.UserRead])
+def read_users(skip: int = 0, limit: int = 10, db: Session = Depends(get_db)):
+    return crud.get_users(db, skip=skip, limit=limit)
+
+@router.put("/users/me", response_model=schemas.UserRead)
+def update_user_profile(
+    updated_user: schemas.UserRegister, 
+    db: Session = Depends(get_db),
+    current_user: models.User = Depends(get_current_user),
+):
+    user = db.query(models.User).filter(models.User.id == current_user.id).first()
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    
+
+    user.name = updated_user.name
+    user.email = updated_user.email
+
+    db.commit()
+    db.refresh(user)
+    return user
+
+@router.delete("/users/me", status_code=status.HTTP_204_NO_CONTENT)
+def delete_user_account(
+    db: Session = Depends(get_db),
+    current_user: models.User = Depends(get_current_user),
+):
+    user = db.query(models.User).filter(models.User.id == current_user.id).first()
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    db.delete(user)
+    db.commit()
+    return
