+++ app/schemas/__init__.py
@@ -1,0 +1,1 @@
+from .user import UserRegister, UserLogin, UserRead, Message, Token
